add prompt controller
